+++ tests/testthat/test-qsip_sample_data.R
@@ -1,12 +1,7 @@
-qsip_sample_data_test_df <- example_sample_df |>
-  add_gradient_pos_rel_amt(source_mat_id = "source", amt = "avg_16S_g_soil") |>
-  dplyr::mutate(Fraction = as.integer(Fraction))
-
-qsip_sample_data_test_df_no_rel <- example_sample_df |>
-  dplyr::mutate(Fraction = as.integer(Fraction))
+add_taxonomy_testdf <- readRDS(test_path("fixtures", "sample_data_rel_amt.rds"))
 
 test_that("works correctly", {
-  expect_snapshot(qsip_sample_data(qsip_sample_data_test_df,
+  expect_snapshot(qsip_sample_data(add_taxonomy_testdf,
     sample_id = "sample",
     source_mat_id = "source",
     gradient_position = "Fraction",
@@ -17,24 +12,27 @@ test_that("works correctly", {
 })
 
 test_that("no gradient_pos_rel_amt given will autocalculate", {
-  expect_snapshot(qsip_sample_data(qsip_sample_data_test_df_no_rel,
-                                   sample_id = "sample",
-                                   source_mat_id = "source",
-                                   gradient_position = "Fraction",
-                                   gradient_pos_density = "density_g_ml",
-                                   gradient_pos_amt = "avg_16S_g_soil"
+  expect_snapshot(qsip_sample_data(example_sample_df,
+    sample_id = "sample",
+    source_mat_id = "source",
+    gradient_position = "Fraction",
+    gradient_pos_density = "density_g_ml",
+    gradient_pos_amt = "avg_16S_g_soil"
   ))
-  expect_message(qsip_sample_data(qsip_sample_data_test_df_no_rel,
-                                   sample_id = "sample",
-                                   source_mat_id = "source",
-                                   gradient_position = "Fraction",
-                                   gradient_pos_density = "density_g_ml",
-                                   gradient_pos_amt = "avg_16S_g_soil"),
-                 "<gradient_pos_rel_amt> not specified. Calculating using avg_16S_g_soil column")
+  expect_message(
+    qsip_sample_data(example_sample_df,
+      sample_id = "sample",
+      source_mat_id = "source",
+      gradient_position = "Fraction",
+      gradient_pos_density = "density_g_ml",
+      gradient_pos_amt = "avg_16S_g_soil"
+    ),
+    "<gradient_pos_rel_amt> not specified. Calculating using avg_16S_g_soil column"
+  )
 })
 
 test_that("missing columns give errors", {
-  expect_error(qsip_sample_data(qsip_sample_data_test_df,
+  expect_error(qsip_sample_data(add_taxonomy_testdf,
     sample_id = "not_found",
     source_mat_id = "source",
     gradient_position = "Fraction",
@@ -42,7 +40,7 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "avg_16S_g_soil",
     gradient_pos_rel_amt = "gradient_pos_rel_amt"
   ), "sample_id column not found")
-  expect_error(qsip_sample_data(qsip_sample_data_test_df,
+  expect_error(qsip_sample_data(add_taxonomy_testdf,
     sample_id = "sample",
     source_mat_id = "not_found",
     gradient_position = "Fraction",
@@ -50,7 +48,7 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "avg_16S_g_soil",
     gradient_pos_rel_amt = "gradient_pos_rel_amt"
   ), "source_mat_id column not found")
-  expect_error(qsip_sample_data(qsip_sample_data_test_df,
+  expect_error(qsip_sample_data(add_taxonomy_testdf,
     sample_id = "sample",
     source_mat_id = "source",
     gradient_position = "not_found",
@@ -58,7 +56,7 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "avg_16S_g_soil",
     gradient_pos_rel_amt = "gradient_pos_rel_amt"
   ), "gradient_position column not found")
-  expect_error(qsip_sample_data(qsip_sample_data_test_df,
+  expect_error(qsip_sample_data(add_taxonomy_testdf,
     sample_id = "sample",
     source_mat_id = "source",
     gradient_position = "Fraction",
@@ -66,7 +64,7 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "avg_16S_g_soil",
     gradient_pos_rel_amt = "gradient_pos_rel_amt"
   ), "gradient_pos_density column not found")
-  expect_error(qsip_sample_data(qsip_sample_data_test_df,
+  expect_error(qsip_sample_data(add_taxonomy_testdf,
     sample_id = "sample",
     source_mat_id = "source",
     gradient_position = "Fraction",
@@ -74,7 +72,7 @@ test_that("missing columns give errors", {
     gradient_pos_amt = "not_found",
     gradient_pos_rel_amt = "gradient_pos_rel_amt"
   ), "gradient_pos_amt column not found")
-  expect_error(qsip_sample_data(qsip_sample_data_test_df,
+  expect_error(qsip_sample_data(add_taxonomy_testdf,
     sample_id = "sample",
     source_mat_id = "source",
     gradient_position = "Fraction",