+++ tests/testthat/test-resample_seed.R
@@ -1,3 +1,10 @@
-test_that("multiplication works", {
-  expect_equal(2 * 2, 4)
+qsip_normal_strict_resampled <- readRDS(test_path("fixtures", "qsip_normal_strict_resampled.rds"))
+qsip_normal_strict_filtered <- readRDS(test_path("fixtures", "qsip_normal_strict_filtered.rds"))
+
+test_that("works correctly", {
+  expect_equal(resample_seed(qsip_normal_strict_resampled), 43)
+})
+
+test_that("object not resampled returns NULL", {
+  expect_null(resample_seed(qsip_normal_strict_filtered))
 })