+++ tests/testthat/test-run_EAF_calculations.R
@@ -1,16 +1,7 @@
-test_qsip = example_qsip_object |>
-  run_feature_filter(  unlabeled_source_mat_ids = get_all_by_isotope(example_qsip_object, "12C"),
-                       labeled_source_mat_ids = c("S178", "S179", "S180"),
-                       min_unlabeled_sources = 6,
-                       min_labeled_sources = 3,
-                       min_unlabeled_fractions = 6,
-                       min_labeled_fractions = 6) |>
-  run_resampling(resamples = 100, with_seed = 44)
-
-
+qsip_normal_strict_resampled <- readRDS(test_path("fixtures", "qsip_normal_strict_resampled.rds"))
 
 test_that("works correctly", {
-  expect_snapshot(run_EAF_calculations(test_qsip))
+  expect_snapshot(run_EAF_calculations(qsip_normal_strict_resampled))
 })
 
 test_that("wrong type or not run through pre-steps", {