+++ tests/testthat/test-run_growth_calculations.R
@@ -1,24 +1,7 @@
-test_growth <- run_feature_filter(example_qsip_growth_object,
-                          group = "Day 10",
-                          unlabeled_source_mat_ids = c("source_11", "source_14", "source_2", "source_5", "source_8"),
-                          labeled_source_mat_ids = c("source_12", "source_15", "source_3", "source_6", "source_9"),
-                          min_labeled_fractions = 4,
-                          min_unlabeled_fractions = 4
-) |>
-  run_resampling(
-    resamples = 100,
-    with_seed = 1332,
-    allow_failures = TRUE,
-    quiet = TRUE
-  ) |>
-  run_EAF_calculations(gc_method = "S", propO = 0.6)
-
-
-
-
+qsip_growth_EAF <- readRDS(test_path("fixtures", "qsip_growth_EAF.rds"))
 
 test_that("works as expected", {
-  expect_snapshot(run_growth_calculations(test_growth,
+  expect_snapshot(run_growth_calculations(qsip_growth_EAF,
                                        N_total_it = example_qsip_growth_t0))
 })
 
@@ -35,12 +18,12 @@ test_that("non-growth qsip object errors", {
 
 
 test_that("Not giving time zero (N_total_it) data failes", {
-  expect_error(run_growth_calculations(test_growth), 'argument "N_total_it" is missing, with no default')
+  expect_error(run_growth_calculations(qsip_growth_EAF), 'argument "N_total_it" is missing, with no default')
 })
 
 
 test_that("designated timepoint column not found gives error", {
-  expect_error(run_growth_calculations(test_growth,
+  expect_error(run_growth_calculations(qsip_growth_EAF,
                                           N_total_it = example_qsip_growth_t0,
                                           timepoint = "not_found"),
                "timepoint column not_found not found in source_data@data")
@@ -48,7 +31,7 @@ test_that("designated timepoint column not found gives error", {
 
 
 test_that("unkonwn growth model fails", {
-  expect_error(run_growth_calculations(test_growth,
+  expect_error(run_growth_calculations(qsip_growth_EAF,
                                        N_total_it = example_qsip_growth_t0,
                                        growth_model = "not_a_model"),
                "growth_model must be either 'exponential' or 'linear', not not_a_model")
@@ -56,7 +39,7 @@ test_that("unkonwn growth model fails", {
 
 
 test_that("unkonwn correct_copy_numbers fails", {
-  expect_error(run_growth_calculations(test_growth,
+  expect_error(run_growth_calculations(qsip_growth_EAF,
                                        N_total_it = example_qsip_growth_t0,
                                        correct_copy_numbers = "not_an_option"),
                "correct_copy_numbers must be either 'filter' or 'adjust', not not_an_option")
@@ -64,7 +47,7 @@ test_that("unkonwn correct_copy_numbers fails", {
 
 
 test_that("unkonwn correct_EAF fails", {
-  expect_error(run_growth_calculations(test_growth,
+  expect_error(run_growth_calculations(qsip_growth_EAF,
                                        N_total_it = example_qsip_growth_t0,
                                        correct_EAF = "not_an_option"),
                "correct_EAF must be either 'filter' or 'adjust', not not_an_option")