+++ NEWS.md
@@ -8,6 +8,7 @@
     * `vignette("resampling")`
 * Added light/dark switch to website documentation
 * Added optional `title` argument to change the title of the plot to `plot_EAF_values()`, `plot_feature_curves()`, `plot_sample_curves()` and `plot_source_wads()`
+* Switched to using `testthat` fixtures
 
 # qSIP2 0.15